requirements.txt
Django~=4.0.2
djangorestframework~=3.13.1
djangorestframework-simplejwt~=5.0.0
redis~=4.3.4
psycopg2~=2.9.6